feat: 컨트롤러 일부 메소드 유틸화
클래스 의존성이 없는 메소드는 유틸함수로 분류하여 관리
