+++ src/controller/EventController.js
@@ -1,14 +1,6 @@
-import { PRICE_OF_MENUS, CATEGORY_OF_MENUS } from '../constants/MenuInfo';
+import controllerUtils from '../utills/eventConstrollerUtils';
 import InputView from '../view/InputView';
 import OutputView from '../view/OutputView';
-import {
-  STATUS,
-  CHRISMAS_D_DAY,
-  GIVEAWAY_EVENT,
-  EVENT,
-  WEEKDAY_EVENT,
-  SPECIAL_EVENT,
-} from '../constants/EventConstants';
 
 class EventController {
   #model;
@@ -33,7 +25,7 @@ class EventController {
   }
 
   setModelAmount() {
-    this.#model.amount = this.calculateTotalAmountBeforeDiscount(
+    this.#model.amount = controllerUtils.calculateTotalAmountBeforeDiscount(
       this.#model.menus
     );
   }
@@ -47,14 +39,14 @@ class EventController {
   }
 
   setModelBenefitAmount() {
-    this.#model.benefitAmount = this.calculateBenefitAmount(
+    this.#model.benefitAmount = controllerUtils.calculateBenefitAmount(
       this.#model.amount,
       this.#model.benefitAmountList
     );
   }
 
   setModelDiscountAmount() {
-    this.#model.discountAmount = this.calculateTotalDiscountAmount(
+    this.#model.discountAmount = controllerUtils.calculateTotalDiscountAmount(
       this.#model.amount,
       this.#model.benefitAmount
     );
@@ -76,15 +68,6 @@ class EventController {
     OutputView.printDecemberEventBadge(this.#model.discountAmount);
   }
 
-  calculateTotalAmountBeforeDiscount(menus) {
-    let amount = 0;
-    menus.forEach((menu) => {
-      amount += Number(PRICE_OF_MENUS[menu.name]) * Number(menu.quantity);
-    });
-
-    return amount;
-  }
-
   calculateBenefitAmountList(amount, date, menus) {
     const benefitAmountList = [
       ['크리스마스 디데이 할인'],
@@ -93,91 +76,16 @@ class EventController {
       ['특별 할인'],
       ['증정 이벤트'],
     ];
-    benefitAmountList[0].push(this.checkChirsmasDdayEvent(date));
-    benefitAmountList[1].push(this.checkWeekdayDiscountEvent(date, menus));
-    benefitAmountList[2].push(this.checkWeekendDiscountEvent(date, menus));
-    benefitAmountList[3].push(this.checkSpecialDiscountEvent(date));
-    benefitAmountList[4].push(this.checkGiveawayEvent(amount));
-    return benefitAmountList;
-  }
-
-  calculateBenefitAmount(amount, benefitAmountList) {
-    const benefitAmount =
-      amount < 10000
-        ? 0
-        : benefitAmountList.reduce((acc, cur) => acc + cur[1], 0);
-
-    return benefitAmount;
-  }
-
-  calculateTotalDiscountAmount(amount, benefitAmount) {
-    if (amount < GIVEAWAY_EVENT.APPLICABLE_AMOUNT) {
-      return benefitAmount ? amount - benefitAmount : amount;
-    }
-    return benefitAmount
-      ? amount - (benefitAmount - GIVEAWAY_EVENT.APPLICABLE_AMOUNT)
-      : amount;
-  }
-
-  checkGiveawayEvent(amount) {
-    if (amount >= GIVEAWAY_EVENT.APPLICABLE_AMOUNT) {
-      return PRICE_OF_MENUS.샴페인;
-    }
-    return STATUS.NON_APPLICABLE_AMOUNT;
-  }
-
-  checkChirsmasDdayEvent(date) {
-    if (date > CHRISMAS_D_DAY.EVENT_DEADLINE) {
-      return STATUS.NON_APPLICABLE_AMOUNT;
-    }
-
-    return (
-      CHRISMAS_D_DAY.DEFAULT_DISCOUNT +
-      CHRISMAS_D_DAY.BONUS_DISCOUNT * (date - 1)
+    benefitAmountList[0].push(controllerUtils.checkChirsmasDdayEvent(date));
+    benefitAmountList[1].push(
+      controllerUtils.checkWeekdayDiscountEvent(date, menus)
     );
-  }
-
-  checkWeekdayDiscountEvent(date, menus) {
-    if (WEEKDAY_EVENT.NON_APPLICABLE_DATES.includes(date)) {
-      return STATUS.NON_APPLICABLE_AMOUNT;
-    }
-
-    const discountAmount = this.calculateDayEventDiscountAmount(menus, [
-      ...CATEGORY_OF_MENUS.DESSERT,
-    ]);
-
-    return discountAmount;
-  }
-
-  checkWeekendDiscountEvent(date, menus) {
-    if (!WEEKDAY_EVENT.NON_APPLICABLE_DATES.includes(date)) {
-      return STATUS.NON_APPLICABLE_AMOUNT;
-    }
-
-    const discountAmount = this.calculateDayEventDiscountAmount(menus, [
-      ...CATEGORY_OF_MENUS.MAIN,
-    ]);
-
-    return discountAmount;
-  }
-
-  checkSpecialDiscountEvent(date) {
-    if (!SPECIAL_EVENT.APPLICABLE_DATES.includes(date)) {
-      return STATUS.NON_APPLICABLE_AMOUNT;
-    }
-    const discountAmount = 1000;
-    return discountAmount;
-  }
-
-  calculateDayEventDiscountAmount(menus, category) {
-    const discountAmount = menus.reduce((amount, menu) => {
-      if (category.includes(menu.name)) {
-        return (amount += EVENT.DISCOUNT * menu.quantity);
-      }
-      return amount;
-    }, 0);
-
-    return discountAmount;
+    benefitAmountList[2].push(
+      controllerUtils.checkWeekendDiscountEvent(date, menus)
+    );
+    benefitAmountList[3].push(controllerUtils.checkSpecialDiscountEvent(date));
+    benefitAmountList[4].push(controllerUtils.checkGiveawayEvent(amount));
+    return benefitAmountList;
   }
 }
